pyproject.toml
[project]
name = "sql-comment-proj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg[binary,pool]>=3.3.3",
    "python-dotenv>=1.2.2",
]
